feat: support different layouts
This commit integrates the layout support from
[`CopilotChat.nvim`][CopilotChat], which some modifications.

There's a new `window` table in the configuration which allows the
layout to be customised.

The job/execution handling is updated a bit to align with this too. Now
we keep a reference to a single buffer around, and track whether it's
visible or hidden etc.

[CopilotChat]: https://github.com/CopilotC-Nvim/CopilotChat.nvim